T-09: close the review's findings
The nine-slice borders were larger than the badge they were drawn on.
BadgeShapeField pinned PixelsPerUnit to the 64px cell, so a cell was one world
unit while BadgeMetrics.Height is 0.4 — the pill's 31px border came to 0.484
units a side against a 0.4-unit badge, and sliced rendering does not shrink a
border below its native size, so the top and bottom slices overlapped and every
rounded rect rendered as a lozenge. PixelsPerUnit is now CellPixels divided by
the badge height, which makes the cell exactly as tall as the badge it is cut
for, and a test holds both borders inside both the height and the narrowest
badge width. The blue badges render with corners now; before they did not.

FontSizeFor(capacity) never read its capacity — the width fit is CellWidth over
one glyph's advance, which is per cell, not per badge. It is BadgeMetrics.FontSize,
a constant, and the test that asserted the old function returned the same value
for eight capacities is gone with it.

WorldBuilder receives a LevelGraph and nothing else again, as #12 wrote it.
Starting power came in as a second parameter; it is now read from the graph's
own preset through MazePreset.Named, the lookup PowerTuning.For and
ContentRecipe.For were already doing by hand on preset.Name.

PowerCeiling moved to Game.Domain, beside PowerBound. It folds the same rule
ActionResolver applies — additive and enemy values add, multipliers multiply —
and a copy of that rule had no business living in presentation. Its saturation
guard stays: stress mints 29 multipliers off a ladder topping out at 4, and
4^29 times the gains overflows a long.

Also from the review: NumberBadge builds its string through the tested
BadgeText.Of instead of concatenating the prefix again on the Unity side;
badge values are int throughout, with long kept only inside the ceiling that
counts digits; WorldBuilder hangs each badge inside the floor loop that already
exists rather than keeping two parallel lists and scanning them; the camera
check is [Conditional("UNITY_EDITOR")] so it compiles out of the player; the
asset lifetime probe moved to its own BadgeAssetCheckCommand and dropped the
ValueTuple and the Union that concatenated; and WorldPreviewCommand's using
directives are back to what T-08 left them as.

Kept, with reasons. LevelBlueprintBuilder.TryProp stays public: the badge needs
the top of the exact prop T-08 raises, and the alternative is a second copy of
that geometry table. The ceiling stays the sound bound rather than a walkable
one — over 300 ship seeds it asks for five glyph cells on 52 of them where four
would do, and the sweep now asserts five rather than six so that cannot drift,
but a player badge that overflows mid-count-up is the failure the criterion
exists to prevent, and no cheap bound is both tighter and provable.

+++ Assets/Scripts/Presentation.Pure/BadgePart.cs
@@ -9,7 +9,7 @@ namespace Game.Presentation.Pure
             int nodeId,
             int floor,
             BadgeStyle style,
-            long value,
+            int value,
             int cells,
             WorldPoint position,
             WorldPoint rotation)
@@ -37,7 +37,7 @@ namespace Game.Presentation.Pure
 
         public BadgeStyle Style { get; }
 
-        public long Value { get; }
+        public int Value { get; }
 
         public int Cells { get; }
